Fix scons git version with submodule. Add helpful error if no source

+++ recompile.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 set -e
 cd `dirname $0`
+
+if [ ! -d sneezymud/code ]; then
+    echo "Please run git submodule update --init"
+    exit 1
+fi
+
+echo 'gitdir: ../.git/modules/sneezymud' > sneezymud/.git
 docker build -t sneezy docker
 docker run -it --rm -v `pwd`:/home/sneezy/sneezymud-docker sneezy
